Update styling mobile form & responsive layout

+++ admin/kategori.php
@@ -51,6 +51,7 @@ $result2 = $stmt2->get_result();
 
 <head>
     <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Kategori <?= htmlspecialchars($kategoriNama) ?></title>
     <link rel="stylesheet" href="../src/output.css">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/7.0.1/css/all.min.css" integrity="sha512-2SwdPD6INVrV/lHTZbO2nodKhrnDdJK9/kg2XD1r9uGqPo1cUbujc+IYdlYdEErWNu69gVcYgdxlmVmzTWnetw==" crossorigin="anonymous" referrerpolicy="no-referrer" />
@@ -59,40 +60,48 @@ $result2 = $stmt2->get_result();
 <body class="">
 
     <!-- Flash Message -->
-    <?php if (isset($_SESSION["flash"])): ?>
-        <div class="fixed bg-green-300 text-green-600 text-center rounded h-7 w-full transition-opacity duration-1000 ease-in-out" id="flash">
-            <?= htmlspecialchars($_SESSION["flash"]) ?>
+    <?php if (isset($_SESSION['flash'])): ?>
+        <div id="flash"
+            class="fixed top-3 left-0 sm:left-1/2 sm:-translate-x-1/2 
+          w-full sm:w-auto sm:max-w-md 
+          bg-emerald-100 text-emerald-800 border border-emerald-300 
+          rounded-md sm:rounded-lg px-4 sm:px-6 py-2 sm:py-3 
+          text-center font-medium text-xs sm:text-sm 
+          shadow-md sm:shadow-lg 
+          z-[9999] animate-slide-down">
+            <?= htmlspecialchars($_SESSION['flash']) ?>
         </div>
-        <?php unset($_SESSION["flash"]); ?>
+        <?php unset($_SESSION['flash']); ?>
     <?php endif; ?>
 
     <!-- Blok Abu -->
     <div class="w-full h-7 bg-gray-200 px-0"></div>
 
     <!-- Navbar -->
-    <nav class="px-8 flex flex-row items-center justify-center gap-4">
+    <nav class="px-1 lg:px-8 flex flex-row items-center justify-center gap-2 lg:gap-4">
 
         <!-- Header -->
         <header class="flex items-center py-4 relative">
-
             <!-- Checkbox trigger -->
             <input type="checkbox" id="menu-toggle" class="hidden peer" />
 
             <!-- Tombol Hamburger -->
             <label for="menu-toggle"
-                class="h-[30px] w-[30px] cursor-pointer flex flex-col items-center justify-center gap-[5px] transition-all duration-300 peer-checked:opacity-0 peer-checked:pointer-events-none z-50">
-                <span class="block h-[3px] w-[30px] bg-black transition-all duration-300"></span>
-                <span class="block h-[3px] w-[30px] bg-black transition-all duration-300"></span>
-                <span class="block h-[3px] w-[30px] bg-black transition-all duration-300"></span>
+                class="h-[30px] w-[30px] cursor-pointer flex flex-col items-center justify-center gap-[4px] lg:gap-[5px] transition-all duration-300 peer-checked:opacity-0 peer-checked:pointer-events-none z-50">
+                <span class="block h-[2px] w-[20px] lg:h-[3px] lg:w-[30px] bg-black transition-all duration-300"></span>
+                <span class="block h-[2px] w-[20px] lg:h-[3px] lg:w-[30px] bg-black transition-all duration-300"></span>
+                <span class="block h-[2px] w-[20px] lg:h-[3px] lg:w-[30px] bg-black transition-all duration-300"></span>
             </label>
 
             <!-- Judul -->
-            <h1 class="ml-4 text-4xl font-lobster">BRIMOB SPORT</h1>
+            <h1 class="hidden sm:block ml-4 mr-2 text-sm lg:text-4xl font-lobster">BRIMOB SPORT</h1>
 
-            <label for="menu-toggle" class="fixed inset-0 bg-transparent hidden peer-checked:block z-30 transition-opacity duration-800 left-64"></label>
+            <label for=" menu-toggle"
+                class="fixed inset-0 bg-transparent hidden peer-checked:block z-30 transition-opacity duration-800 left-64"></label>
 
             <!-- Sidebar BRIMOB SPORT -->
-            <div id="sidebar" class="fixed top-0 left-0 h-screen w-64 bg-white border-r border-gray-200 shadow-[4px_0_15px_rgba(0,0,0,0.05)] transition-transform duration-300 -translate-x-full peer-checked:translate-x-0 z-20 flex flex-col font-inter">
+            <div id="sidebar"
+                class="fixed top-0 left-0 h-screen w-64 bg-white border-r border-gray-200 shadow-[4px_0_15px_rgba(0,0,0,0.05)] transition-transform duration-300 -translate-x-full peer-checked:translate-x-0 z-20 flex flex-col font-inter">
 
                 <!-- Logo judul sidebar -->
                 <div class="px-6 py-8 border-b border-gray-200 text-center">
@@ -157,9 +166,11 @@ $result2 = $stmt2->get_result();
 
         <!-- Form search -->
         <form action="../produk/search.php" method="get" class="relative flex-1 group">
+            <input type="hidden" name="kategori" value="<?= urlencode(strtolower($kategoriNama)) ?>">
+            <input type="hidden" name="from" value="kategori">
             <!-- Ikon search -->
-            <span class="absolute inset-y-0 left-0 flex items-center pl-3 text-gray-400 transition-all duration-300 group-focus-within:text-black">
-                <svg class="w-5 h-5" xmlns="http://www.w3.org/2000/svg" fill="none"
+            <span class="hidden absolute inset-y-0 left-0 sm:flex items-center pl-[2px] lg:pl-3 text-gray-400 transition-all duration-300 group-focus-within:text-black">
+                <svg class="w-3 h-3  lg:w-5 lg:h-5" xmlns="http://www.w3.org/2000/svg" fill="none"
                     viewBox="0 0 24 24" stroke="currentColor">
                     <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
                         d="M21 21l-4.35-4.35m0 0A7.5 7.5 0 1110.5 3a7.5 7.5 0 016.15 13.65z" />
@@ -168,34 +179,38 @@ $result2 = $stmt2->get_result();
 
             <!-- Input -->
             <input type="text" name="q" placeholder="Cari produk keren di BRIMOB SPORT..."
-                class="w-full pl-10 pr-20 py-2 h-10 border border-gray-400 rounded-lg bg-gradient-to-r from-gray-50 to-white text-gray-800 placeholder-gray-400
-                        focus:outline-none focus:border-black focus:ring-2 focus:ring-gray-700 focus:shadow-[0_0_12px_rgba(0,0,0,0.15)] transition-all duration-300 ease-in-out shadow-sm
-                        hover:shadow-[0_0_10px_rgba(0,0,0,0.1)]" />
+                class="h-6 rounded-sm w-full pl-[2px] sm:pl-5 lg:pl-10 pr-[20px] lg:pr-20 leading-[1.1rem] lg:leading-[1.25rem] lg:h-10 border border-gray-400 lg:rounded-lg bg-gradient-to-r from-gray-50 to-white text-gray-800 placeholder-gray-400 placeholder:py-1
+                focus:outline-none focus:border-black focus:ring-2 focus:ring-gray-700 focus:shadow-[0_0_12px_rgba(0,0,0,0.15)] transition-all duration-300 ease-in-out 
+                lg:text-base placeholder:text-xs sm:placeholder:text-sm lg:placeholder:text-base shadow-sm hover:shadow-[0_0_10px_rgba(0,0,0,0.1)]" />
 
             <!-- Tombol -->
-            <button type="submit"
-                class="absolute right-0 top-0 h-10 px-4 bg-black text-white rounded-r-lg hover:bg-gray-900 active:scale-[0.97] font-medium transition-all duration-200 ease-in-out cursor-pointer
-                        shadow-[0_2px_6px_rgba(0,0,0,0.2)] hover:shadow-[0_2px_10px_rgba(0,0,0,0.25)]">
+            <button
+                type="submit"
+                class="absolute right-0 top-0 sm:top-[3.5px] lg:text-[16px] lg:top-0 h-6 sm:h-[17.5px] lg:h-10 px-[5px] sm:px-[7px] lg:px-4 bg-black text-white text-[8.5px] sm:text-[10px] rounded-r-sm lg:rounded-r-lg flex items-center justify-center hover:bg-gray-900 active:scale-[0.97] font-medium transition-all duration-200 ease-in-out cursor-pointer shadow-[0_1px_2px_rgba(0,0,0,0.2)] hover:shadow-[0_2px_4px_rgba(0,0,0,0.25)]">
                 Search
             </button>
         </form>
 
         <!-- Jika admin login -->
         <?php if (!empty($_SESSION['admin']['foto'])): ?>
+            <div class="w-[2px] h-5 lg:w-1 lg:h-10 bg-gray-300"></div>
             <a href="edit_profil_admin.php?from=kategori&kategori=<?= urlencode($_GET['nama']) ?>" class="flex items-center gap-4 text-2xl">
                 <img src="../<?= htmlspecialchars($_SESSION['admin']['foto']) ?>"
                     alt="Foto Admin"
-                    class="size-10 rounded-full object-cover">
+                    class="size-6 sm:size-8 lg:size-10 rounded-full object-cover">
             </a>
         <?php else: ?>
             <a href="edit_profil_admin.php?from=kategori&kategori=<?= urlencode($_GET['nama']) ?>" class="flex items-center gap-4 text-2xl">
-                <i class="fas fa-user-circle text-3xl"></i>
+                <div
+                    class="flex items-center justify-center w-8 h-8 sm:w-9 sm:h-9 lg:w-10 lg:h-10 rounded-full bg-gray-200 border border-gray-300">
+                    <i class="fas fa-user text-gray-600 text-[18px] sm:text-[20px] lg:text-[22px]"></i>
+                </div>
             </a>
         <?php endif; ?>
 
         <!-- Logout -->
         <button class="flex items-center justify-center cursor-pointer" id="openModal" type="button">
-            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="size-10">
+            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="size-6 lg:size-10">
                 <path stroke-linecap="round" stroke-linejoin="round" d="M15.75 9V5.25A2.25 2.25 0 0 0 13.5 3h-6a2.25 2.25 0 0 0-2.25 2.25v13.5A2.25 2.25 0 0 0 7.5 21h6a2.25 2.25 0 0 0 2.25-2.25V15m3 0 3-3m0 0-3-3m3 3H9" />
             </svg>
         </button>
@@ -204,13 +219,13 @@ $result2 = $stmt2->get_result();
     <hr class=" border-gray-500 mb-8">
 
     <!-- Kategori Produk -->
-    <main class="px-30">
+    <main class="px-8 lg:px-30">
         <div>
-            <h1 class="text-3xl font-bold mb-6">Categories: <?= htmlspecialchars($kategoriNama) ?></h1>
+            <h1 class="font-bold text-2xl lg:text-3xl font-inter">Categories: <?= htmlspecialchars($kategoriNama) ?></h1>
             <br>
 
             <!-- Tombol kategori -->
-            <div class="flex flex-row gap-10">
+            <div class="grid grid-cols-1 gap-5 lg:flex lg:flex-row lg:gap-10">
                 <?php
                 $result = $koneksi->query("SELECT nama, slug FROM kategori");
                 while ($row = $result->fetch_assoc()):
@@ -227,48 +242,53 @@ $result2 = $stmt2->get_result();
 
             <br>
 
-            <!-- Prosuk -->
-            <div class="grid grid-cols-4 gap-10">
+            <!-- Produk -->
+            <div class="grid grid-cols-3 gap-3 lg:grid-cols-4 lg:gap-10">
                 <?php while ($row = $result2->fetch_assoc()): ?>
-                    <div class="px-2 py-4 h-[320px] w-2xs rounded-2xl flex flex-col justify-between items-center shadow-[0_0_10px_rgba(0,0,0,0.3)] hover:scale-105 transition-transform duration-200 ease-in-out cursor-pointer">
+                    <div class="w-auto sm:w-[180px] lg:px-2 py-4 h-auto lg:w-auto rounded-2xl flex flex-col justify-between items-center shadow-[0_0_10px_rgba(0,0,0,0.3)] hover:scale-105 transition-transform duration-200 ease-in-out cursor-pointer">
                         <a href="../produk/detail_produk.php?id=<?= $row['id'] ?>&from=kategori&kategori=<?= urlencode($slug) ?>" class="flex flex-col justify-between items-center">
                             <!-- gambar -->
                             <img src="../img/<?= htmlspecialchars($row['gambar']) ?>" alt="<?= htmlspecialchars($row['nama']) ?>"
-                                class="w-40 h-40 object-contain mb-4">
+                                class="h-20 w-40 px-2 lg:h-40 object-contain mb-4">
+                            <!-- Nama -->
+                            <p class="font-bold px-[8px] text-[10px] text-center lg:text-lg lg:text-center min-h-[60px]">
+                                <?= $row['nama'] ?>
+                            </p>
 
-                            <!-- nama -->
-                            <p class="font-bold text-xl text-center min-h-[60px]"><?= htmlspecialchars($row['nama']) ?></p>
-
-                            <!-- harga -->
-                            <p class="text-lg text-gray-600"><?= "Rp " . number_format($row['harga'], 2, ',', '.') ?></p>
+                            <!-- Harga -->
+                            <p class="text-[8px] font-bold lg:text-lg text-gray-600">
+                                <?= "Rp " . number_format($row['harga'], 2, ',', '.') ?>
+                            </p>
                         </a>
 
                         <!-- Tombol aksi -->
-                        <div class="flex gap-2 mt-2">
-                            <a href="edit_produk.php?id=<?= $row['id'] ?>&from=kategori&kategori=<?= urlencode($slug) ?>"
-                                class="px-3 py-1 bg-blue-500 text-white rounded hover:bg-blue-600 text-sm">Edit</a>
-                            <a href="hapus_produk.php?id=<?= $row['id'] ?>&from=kategori&kategori=<?= urlencode($slug) ?>"
-                                onclick="return confirm('Yakin ingin menghapus produk ini?')"
-                                class="px-3 py-1 bg-red-500 text-white rounded hover:bg-red-600 text-sm">Hapus</a>
-                        </div>
+                        <?php if (isset($_SESSION['admin'])): ?>
+                            <div class="cursor-pointer flex items-start justify-between lg:justify-evenly gap-1 sm:gap-2 w-full mt-2 sm:mt-3 lg:px-2 px-1">
+                                <a href="edit_produk.php?id=<?= $row['id'] ?>&from=kategori&kategori=<?= urlencode($slug) ?>"
+                                    class="w-12 py-[6px] px-2 h-6 sm:w-16 sm:h-10 text-center text-[7px] sm:text-[16px] font-medium bg-blue-500 text-white rounded hover:bg-blue-600 text-sm">Edit</a>
+                                <a href="hapus_produk.php?id=<?= $row['id'] ?>&from=kategori&kategori=<?= urlencode($slug) ?>"
+                                    onclick="return confirm('Yakin ingin menghapus produk ini?')"
+                                    class="w-12 py-[6px] px-2 h-6 sm:w-16 sm:h-10 text-center text-[7px] sm:text-[16px] font-medium bg-red-500 text-white rounded hover:bg-red-600 text-sm">Hapus</a>
+                            </div>
+                        <?php endif; ?>
                     </div>
                 <?php endwhile; ?>
 
                 <!-- Tombol tambah produk -->
                 <a href="tambah_produk.php?from=kategori&kategori=<?= urlencode($slug) ?>"
-                    class="h-[320px] w-2xs rounded-2xl flex flex-col justify-center items-center shadow-[0_0_10px_rgba(0,0,0,0.3)] hover:scale-105 transition-transform duration-200 ease-in-out cursor-pointer">
-                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="size-60">
+                    class="w-auto sm:w-[180px] lg:px-2 py-4 h-auto lg:w-auto rounded-2xl flex flex-col justify-between items-center shadow-[0_0_10px_rgba(0,0,0,0.3)] hover:scale-105 transition-transform duration-200 ease-in-out cursor-pointer">
+                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="size-20 lg:size-60">
                         <path stroke-linecap="round" stroke-linejoin="round" d="M12 9v6m3-3H9m12 0a9 9 0 1 1-18 0 9 9 0 0 1 18 0Z" />
                     </svg>
-                    <span class="text-2xl font-bold">Tambah Produk</span>
+                    <span class="text-sm text-center lg:text-2xl font-bold">Tambah Produk</span>
                 </a>
             </div>
         </div>
         </div>
 
         <!-- Kembali -->
-        <div class="mt-8">
-            <a href="dashboard_admin.php" class="px-6 py-3 bg-gray-500 text-white rounded-lg hover:bg-gray-600">← Kembali</a>
+        <div class="mt-8 mb-8">
+            <a href="dashboard_admin.php" class="px-2 lg:px-6 py-3 bg-gray-500 text-white rounded-lg hover:bg-gray-600">← Kembali</a>
         </div>
     </main>
